requirements.txt
boto3==1.21.13
loguru==0.6.0
python-dotenv==0.19.2
slack-sdk==3.15.2

